Added model that allows bootstrapped values be sent to some children only.
This also required some changes to the input processor. Dependencies in Lingo are no longer ordered (child, parent), but rather (parent, child), which seems to be more common and natural, and fits with our formal description of the model. The solution validator has not yet been updated to validate output from this new model.