Fixed redundant path removal for selective model.
The custom_ddgs_format_parser files incorrectly removed paths from the bootstrapping paths list when using the selective model. The larger_path_contains_smaller_path function has been added and is called in the path_is_redundant function so that whether a path is redundant is handled differently for the selective model and other models.